# Speed up arbtt-stats --filter some more