This is pretty silly, I know. I should just implement the logfile rotation instead. But it's so amusing that I can't resist sharing, though. :-)

While fine-tuning my categorizer.cfg I noticed that arbtt-stats is significantly faster with `--also-inactive` and a date filter, and it turns out that this is because `--also-inactive` means we only need to run the full categorizer for the samples matching the date filter, whereas normally we'd categorize everything, check for the inactive tag, and only then apply the date filter. This is confirmed by measuring arbtt-stats with `--also-inactive --exclude=inactive --filter=…`, which applies the filter first and the exclude after, and it is indeed more than twice as fast.

So this commit adds a heuristic that reorders filters to always apply `--filter` first and `--exclude`/`--only` last, including the implicit `--exclude=inactive`. As long as the command-line filters are cheaper than `categorize.cfg`, this should result in a speedup.

Benchmarks on my small alternate X server logfile follow. With my primary logfile, this brings arbtt-stats from 9 seconds down to a little over 3.

before:

    $ hyperfine …
    Benchmark #1: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log
      Time (mean ± σ):      76.5 ms ±   6.6 ms    [User: 70.2 ms, System: 6.3 ms]
      Range (min … max):    73.3 ms … 107.9 ms    36 runs

    Benchmark #2: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --filter="\$sampleage < 24:00"
      Time (mean ± σ):      66.2 ms ±   1.3 ms    [User: 59.9 ms, System: 6.4 ms]
      Range (min … max):    63.5 ms …  70.3 ms    43 runs

    Benchmark #3: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --also-inactive --filter="\$sampleage < 24:00"
      Time (mean ± σ):      32.9 ms ±   0.7 ms    [User: 26.2 ms, System: 6.6 ms]
      Range (min … max):    31.7 ms …  36.4 ms    87 runs

    Benchmark #4: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --also-inactive --exclude=inactive --filter="\$sampleage < 24:00"
      Time (mean ± σ):      33.2 ms ±   1.1 ms    [User: 27.1 ms, System: 6.0 ms]
      Range (min … max):    31.7 ms …  37.1 ms    80 runs

    Benchmark #5: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --also-inactive --filter="\$sampleage < 24:00" --exclude=inactive
      Time (mean ± σ):      66.4 ms ±   1.5 ms    [User: 59.2 ms, System: 7.1 ms]
      Range (min … max):    64.2 ms …  73.8 ms    44 runs

after:

    $ hyperfine …
    Benchmark #1: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log
      Time (mean ± σ):      74.9 ms ±   1.2 ms    [User: 68.0 ms, System: 6.7 ms]
      Range (min … max):    73.0 ms …  77.6 ms    38 runs

    Benchmark #2: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --filter="\$sampleage < 24:00"
      Time (mean ± σ):      33.0 ms ±   1.0 ms    [User: 26.5 ms, System: 6.3 ms]
      Range (min … max):    31.8 ms …  39.4 ms    87 runs

    Benchmark #3: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --also-inactive --filter="\$sampleage < 24:00"
      Time (mean ± σ):      32.9 ms ±   1.2 ms    [User: 26.6 ms, System: 6.2 ms]
      Range (min … max):    31.9 ms …  41.2 ms    87 runs

    Benchmark #4: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --also-inactive --exclude=inactive --filter="\$sampleage < 24:00"
      Time (mean ± σ):      32.9 ms ±   0.7 ms    [User: 25.9 ms, System: 7.0 ms]
      Range (min … max):    31.9 ms …  36.3 ms    80 runs

    Benchmark #5: arbtt-stats --logfile=/home/tomi/.arbtt/capture-:1.log --also-inactive --filter="\$sampleage < 24:00" --exclude=inactive
      Time (mean ± σ):      33.1 ms ±   1.7 ms    [User: 27.1 ms, System: 5.9 ms]
      Range (min … max):    31.7 ms …  45.0 ms    81 runs